chore: inline dcos-ui-update-service-client
reorganizing our code in order to get rid of a couple of "plugin"-mechanisms.

/packages looks like the easiest win.
