--- packages/dcos-ui-update-client/index.ts
@@ -1,7 +1,0 @@
-export { DCOSUIUpdateClient } from "./src/DCOSUIUpdateClient";
-
-export interface UIVersionResponse {
-  default: boolean;
-  packageVersion: string;
-  buildVersion: string;
-}
