--- packages/dcos-ui-update-client/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "dcos-ui-update-client",
-  "module": "index.ts",
-  "version": "1.0.0",
-  "license": "Apache-2.0",
-  "scripts": {},
-  "dependencies": {
-    "@dcos/http-service": "2.3.2"
-  },
-  "peerDependencies": {
-    "rxjs": "^6.3.3"
-  },
-  "devDependencies": {
-    "@types/jest": "22.2.3",
-    "jest-cli": "22.4.0"
-  },
-  "jest-webpack-alias": {
-    "configFile": "../../webpack/webpack.config.js"
-  }
-}
