+++ src/js/data/ui-update/DCOSUIUpdateClient.ts
@@ -2,9 +2,11 @@ import { Observable } from "rxjs";
 import { tap } from "rxjs/operators";
 import { request, RequestResponse } from "@dcos/http-service";
 
-import { handleBadResponses } from "./utils";
-
-import { UIVersionResponse } from "../";
+export interface UIVersionResponse {
+  default: boolean;
+  packageVersion: string;
+  buildVersion: string;
+}
 
 export const DCOSUIUpdateClient = {
   fetchVersion: (
@@ -33,3 +35,20 @@ export const DCOSUIUpdateClient = {
     }).pipe(tap(handleBadResponses));
   }
 };
+
+export function getErrorMessage(reqResp: RequestResponse<any>): string {
+  const { response, message } = reqResp;
+  if (typeof response === "string") {
+    return response;
+  }
+  if (message) {
+    return message;
+  }
+  return "An error occurred.";
+}
+
+export function handleBadResponses(requestResponse: RequestResponse<any>) {
+  if (requestResponse.code >= 300) {
+    throw new Error(getErrorMessage(requestResponse));
+  }
+}
