+++ src/js/data/ui-update/__tests__/DCOSUIUpdateClient-test.ts
